// swiyu-ecosystem-portal-ui/src/app/api/generated/model/protected-verification-submission-list-item.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface ProtectedVerificationSubmissionListItem {
  id: string;
  partnerId: string;
  entityName: string;
  category: ProtectedVerificationSubmissionListItem.CategoryEnum;
  status: ProtectedVerificationSubmissionListItem.StatusEnum;
  submittedAt: string;
  createdAt: string;
  updatedAt: string;
}
export namespace ProtectedVerificationSubmissionListItem {
  export const CategoryEnum = {
    PersonalAdministrativeNumber: 'PERSONAL_ADMINISTRATIVE_NUMBER'
  } as const;
  export type CategoryEnum = (typeof CategoryEnum)[keyof typeof CategoryEnum];
  export const StatusEnum = {
    Submitted: 'SUBMITTED',
    Approved: 'APPROVED',
    Rejected: 'REJECTED'
  } as const;
  export type StatusEnum = (typeof StatusEnum)[keyof typeof StatusEnum];
}